Fix regression in 64-bit in r882 i_mvc needs to be 64-bit when used with a 64-bit memory pointer

+++ common/common.h
@@ -130,7 +130,7 @@ static inline void x264_median_mv( int16_t *dst, int16_t *a, int16_t *b, int16_t
     dst[1] = x264_median( a[1], b[1], c[1] );
 }
 
-static inline int x264_predictor_difference( int16_t (*mvc)[2], int i_mvc )
+static inline int x264_predictor_difference( int16_t (*mvc)[2], intptr_t i_mvc )
 {
     int sum = 0, i;
     for( i = 0; i < i_mvc-1; i++ )
